TECH: Refactor map hooks and tools
