--- src/react/features/lenses/map/hooks/index.js
@@ -1,2 +1,0 @@
-export { useMap } from './useMap'
-export { useMapMarkers } from './useMapMarkers'
